requirements-dev.txt
fastapi==0.115.6
uvicorn==0.34.0
chromadb==0.6.3
crewai==0.95.0
langchain-openai==0.3.1
langchain-community==0.3.15
python-docx==1.1.2
python-dotenv==1.0.1
pydantic==2.10.5
pytest==8.3.4
loguru==0.7.3
tenacity==9.0.0
numpy==1.26.4
openai==1.59.7
unstructured==0.17.2
unstructured-inference==0.8.10
unstructured_pytesseract==0.3.15
gmft==0.4.0
PyMuPDF==1.25.1
tqdm==4.67.1
Pillow==11.1.0
python-multipart==0.0.20
pi_heif==0.21.0
APScheduler==3.11.0
pdf2image==1.17.0
accelerate==1.6.0
pika==1.3.2